Fix a really annoying problem introduced by including cstdlib: This bought us an ambiguity between abs(int) and abs(long int) when calling it with the difference between two unsigned integers... Sometimes C++ is less than only C.
git-svn-id: https://svn.dealii.org/trunk@789 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/source/dsmatrix.cc
@@ -346,8 +346,8 @@ dSMatrixStruct::bandwidth () const
     for (unsigned int j=rowstart[i]; j<rowstart[i+1]; ++j)
       if (colnums[j]>=0) 
 	{
-	  if ((unsigned int)abs(i-colnums[j]) > b)
-	    b = abs(i-colnums[j]);
+	  if (static_cast<unsigned int>(abs(static_cast<int>(i-colnums[j]))) > b)
+	    b = abs(static_cast<int>(i-colnums[j]));
 	}
       else
 					 // leave if at the end of
